fix: Select icon alignment
