+++ packages/components/src/Form/Inputs/Select/Select.test.tsx
@@ -971,6 +971,40 @@ describe('Select', () => {
     fireEvent.click(document)
   })
 
+  test('option overline', () => {
+    renderWithTheme(
+      <Select
+        placeholder="Search"
+        options={[
+          { label: 'Foo', overline: 'Overline for Foo', value: 'FOO' },
+          {
+            label: 'Bar',
+            overline: (
+              <>
+                <strong>Overline for</strong> Bar
+              </>
+            ),
+            value: 'BAR',
+          },
+        ]}
+      />
+    )
+
+    const input = screen.getByPlaceholderText('Search')
+    fireEvent.click(input)
+    const fooOverline = screen.getByText('Overline for Foo')
+    const barOverline = screen.getByText('Overline for')
+    expect(fooOverline).toBeVisible()
+    expect(barOverline).toBeVisible()
+    expect(barOverline.parentElement).toHaveTextContent('Overline for Bar')
+
+    fireEvent.click(fooOverline)
+    expect(input).toHaveDisplayValue('Foo')
+
+    // Close popover to silence act() warning
+    fireEvent.click(document)
+  })
+
   test('filtering after selecting an option where value != label', () => {
     const customLabelOptions = [
       { label: 'Foo', value: 'FOO' },
