+++ packages/components/src/Form/Inputs/Select/stories/SelectMulti.story.tsx
@@ -26,9 +26,9 @@
 
 import { Story } from '@storybook/react/types-6-0'
 import React from 'react'
-import { Select, SelectProps } from '../../Inputs/Select'
+import { SelectMulti, SelectMultiProps } from '../SelectMulti'
 
-const Template: Story<SelectProps> = (args) => <Select {...args} />
+const Template: Story<SelectMultiProps> = (args) => <SelectMulti {...args} />
 
 export const Basic = Template.bind({})
 Basic.args = {
@@ -43,18 +43,6 @@ export const Placeholder = Template.bind({})
 Placeholder.args = {
   ...Basic.args,
   placeholder: 'Placeholder',
-}
-
-export const AutoResize = Template.bind({})
-AutoResize.args = {
-  ...Basic.args,
-  autoResize: true,
-}
-
-export const AutoResizePlaceholder = Template.bind({})
-AutoResizePlaceholder.args = {
-  ...Placeholder.args,
-  autoResize: true,
 }
 
 export const Disabled = Template.bind({})
@@ -81,53 +69,39 @@ ErrorPlaceholder.args = {
   validationType: 'error',
 }
 
-export const Value = Template.bind({})
-Value.args = {
+export const Values = Template.bind({})
+Values.args = {
   ...Basic.args,
-  value: 'gouda',
+  values: ['cheddar', 'gouda'],
 }
 
-export const ErrorValue = Template.bind({})
-ErrorValue.args = {
-  ...Value.args,
+export const DefaultValues = Template.bind({})
+DefaultValues.args = {
+  ...Basic.args,
+  defaultValues: ['gouda', 'swiss'],
+}
+
+export const WrappingValues = Template.bind({})
+WrappingValues.args = {
+  ...Basic.args,
+  defaultValues: ['cheddar', 'gouda', 'swiss'],
+  width: 250,
+}
+
+export const ErrorValues = Template.bind({})
+ErrorValues.args = {
+  ...Error.args,
+  values: ['cheddar', 'gouda'],
+}
+
+export const ErrorWrappingValues = Template.bind({})
+ErrorWrappingValues.args = {
+  ...WrappingValues.args,
+  defaultValues: ['cheddar', 'gouda', 'swiss'],
   validationType: 'error',
-}
-
-export const DefaultValue = Template.bind({})
-DefaultValue.args = {
-  ...Basic.args,
-  defaultValue: 'swiss',
-}
-
-export const Clearable = Template.bind({})
-Clearable.args = {
-  ...Basic.args,
-  isClearable: true,
-}
-
-export const ClearableValue = Template.bind({})
-ClearableValue.args = {
-  ...Clearable.args,
-  value: 'gouda',
-}
-
-export const ClearableError = Template.bind({})
-ClearableError.args = {
-  ...ClearableValue.args,
-  validationType: 'error',
-}
-
-export const Icon = Template.bind({})
-Icon.args = {
-  options: [
-    { icon: 'ChartBar', label: 'Bar', value: 'bar' },
-    { icon: 'ChartPie', label: 'Pie', value: 'pie' },
-    { icon: 'ChartTable', label: 'Table', value: 'table' },
-  ],
-  value: 'pie',
 }
 
 export default {
-  component: Select,
-  title: 'Select',
+  component: SelectMulti,
+  title: 'SelectMulti',
 }
